requirements.txt
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
websockets==12.0
schedule==1.2.0
colorama==0.4.6
tabulate==0.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3